Deduplicate SetTooltipEntrySettings
